use common

+++ audio/cWinAudio16.cpp
@@ -6,7 +6,7 @@
 #include "cWinAudio16.h"
 
 #include <cstdint>
-#include "../utils/cLog.h"
+#include "../common/cLog.h"
 
 #pragma comment(lib,"Xaudio2.lib")
 //}}}
@@ -146,7 +146,7 @@ void cAudio::play (int srcChannels, void* srcSamples, int srcNumSamples, float p
             }
             //}}}
           }
-        cLog::log (LOGNOTICE, fmt::format ("6 to 2 mixdown changed to {}",mMixDown));
+        cLog::log (LOGNOTICE, fmt::format ("6 to 2 mixdown changed to {}", (int)mMixDown));
         }
         //}}}
       else if (mDstChannels == 4) {
@@ -293,7 +293,7 @@ void cAudio::play (int srcChannels, void* srcSamples, int srcNumSamples, float p
             }
             //}}}
           }
-        cLog::log (LOGNOTICE, fmt::format ("6 to 6 mixdown changed to {}",mMixDown));
+        cLog::log (LOGNOTICE, fmt::format ("6 to 6 mixdown changed to {}", (int)mMixDown));
         }
         //}}}
       }
@@ -305,7 +305,7 @@ void cAudio::play (int srcChannels, void* srcSamples, int srcNumSamples, float p
                                   1.f, 0.f,  // dst L
                                   0.f, 1.f}; // dst R
         mSourceVoice->SetOutputMatrix (mMasteringVoice, mSrcChannels, mDstChannels, kLevelMatrix, XAUDIO2_COMMIT_NOW);
-        cLog::log (LOGNOTICE, fmt::format("2 to 2 mixdown changed to{}  nothing changed ", mMixDown));
+        cLog::log (LOGNOTICE, fmt::format("2 to 2 mixdown changed to{}  nothing changed ", (int)mMixDown));
         }
         //}}}
       else if (mDstChannels == 4) {
@@ -328,7 +328,7 @@ void cAudio::play (int srcChannels, void* srcSamples, int srcNumSamples, float p
                                    1.f, 0.f,  // dst BL
                                    0.f, 1.f}; // dst BR
         mSourceVoice->SetOutputMatrix (mMasteringVoice, mSrcChannels, mDstChannels, kLevelMatrix, XAUDIO2_COMMIT_NOW);
-        cLog::log (LOGNOTICE, fmt::format ("2 to 6 mixdown changed to {} nothing changed",mMixDown));
+        cLog::log (LOGNOTICE, fmt::format ("2 to 6 mixdown changed to {} nothing changed", (int)mMixDown));
         }
         //}}}
       }
